+++ src/app/app.component.ts
@@ -4,11 +4,17 @@
  * For license information see LICENSE file.
  */
 
-import { AuthenticationService, AuthorizationService, SnackbarComponent } from '@abraxas/base-components';
+import {
+  AuthenticationService,
+  AuthorizationService,
+  CornerRadiusTokensThemes,
+  SnackbarComponent,
+  StylingService,
+} from '@abraxas/base-components';
 import { OAuthService } from 'angular-oauth2-oidc';
-import { SnackbarService, ThemeService, LanguageService } from '@abraxas/voting-lib';
+import { LanguageService, SnackbarService, ThemeService } from '@abraxas/voting-lib';
 import { LocationStrategy } from '@angular/common';
-import { Component, HostBinding, OnDestroy, OnInit, ViewChild, inject } from '@angular/core';
+import { Component, HostBinding, inject, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import moment from 'moment';
 import 'moment/locale/de';
@@ -49,6 +55,9 @@ export class AppComponent implements OnInit, OnDestroy {
   constructor() {
     const cursorService = inject(CursorService);
     const themeService = inject(ThemeService);
+    const stylingService = inject(StylingService);
+
+    stylingService.setRadius(CornerRadiusTokensThemes.Default);
 
     // enable automatic silent refresh
     this.oauthService.setupAutomaticSilentRefresh({}, 'access_token');