+++ src/app/majority-elections/majority-election-candidates/majority-election-candidates.component.ts
@@ -30,7 +30,6 @@ import { DomainOfInfluenceParty } from '../../core/models/domain-of-influence-pa
 @Component({
   selector: 'app-majority-election-candidates',
   templateUrl: './majority-election-candidates.component.html',
-  styleUrls: ['./majority-election-candidates.component.scss'],
   standalone: false,
 })
 export class MajorityElectionCandidatesComponent {