+++ src/app/shared/political-business-union-detail-political-businesses-tab/political-business-union-detail-political-businesses-tab.component.ts
@@ -14,6 +14,7 @@ import { PaginatorComponent, TableDataSource } from '@abraxas/base-components';
 @Component({
   selector: 'app-political-business-union-detail-political-businesses-tab',
   templateUrl: './political-business-union-detail-political-businesses-tab.component.html',
+  styleUrl: './political-business-union-detail-political-businesses-tab.component.scss',
   standalone: false,
 })
 export class PoliticalBusinessUnionDetailPoliticalBusinessesTabComponent implements AfterViewInit {